System Hang on Boot Fix
If you have the meta domain blocker on or a restricted internet connection, the system can hang on boot. This is due to the system seeing a wifi connection, and trying its best to connect to certain domains. This loops for a while until it eventually gives up and runs like normal. This should help with that issue :)