Domain Blocker Upgrade
This is a two part system. Non rooted users have a kill switch (Old VPN domain blocker), with that enabled, nothing will get through. This is so that unrooted/patched users still have a potential way to block metas domains. This sadly is the only option for unrooted/patched users. Local VPNs set to block certain domains does not work on the Quest. Its not possible.

The second system in place is a root method that bind mounts the systems host file, applying the blocked domains. This happens on boot.

Enabling the kill switch on boot will also enable the root domain blocker method on "boot". AKA when root is acquired.