Faster UI Updating
This does 2 things:
Tweaksactivity already has what the buttons should be set to saved and cached. When opening the activity, it uses that as the "truth" and checks the states of the toggles in the background to compare it to the "truth". If the check comes back with a different response, then it updates the saved state and updates that in the UI